Reorganize FilterModel to prevent cyclic dependency and enable to file imports
